requirements.txt
anyio==3.6.2
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
fastapi==0.89.1
h11==0.14.0
idna==3.4
numpy==1.24.1
pandas==1.5.2
peewee==3.15.4
pycparser==2.21
pydantic==1.10.4
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.7
pytz==2022.7.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0